Exercicios1_24/Exercise_12/1.app_funtion_azurite_Insert_BlobFile&BD_AZURE/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
requests    
azure-storage-blob
uuid


